+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this programmer does her best to follow [Conventional Commits](https://www.c
 - Relevent states added to GameState
 - Added `Vector2.__iter__`, rendering `tuple(v)`, `*v`, `x, y = v` available
 - Added name to boid class
+- Added save state on key input => o
 
 ### Changed